pyproject.toml
[project]
name = "yolov9-mlx"
version = "0.1.0"
description = "Yolov9 model in MLX"
authors = [
    {name = "SuperBo", email = "supernbo@gmail.com"},
]
dependencies = [
    "mlx>=0.5",
    "numpy>=1.25",
]
requires-python = ">=3.11"
readme = "README.md"
license = {text = "MIT"}

[project.optional-dependencies]
serving = [
    "bentoml>=1.2.0",
    "pillow>=10.0.0",
]

[build-system]
requires = ["setuptools>=61", "wheel"]
build-backend = "setuptools.build_meta"

[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
test = [
    "pytest>=7.4",
]
convert = [
    "torch>=2.1.0",
]

[tool.pdm.scripts]
convert = {call = "yolov9_mlx.cli.convert:main"}

[tool.pyright]
include = ["src"]
exclude = [".venv"]
venvPath = "."
venv = ".venv"
